Improved SSH repositories, more functionalities implemented
There is definitely a need to re-implement repository functions when the repository is in SSH, particularily the part responsible for accesing changes. This is partly done, but may still be enhanced.
Easyproject was updated to function better, and now basic scenario seems to work well, and was even tested with Easytools.
[TAGS] WIP