Added easychanges, updated easyapi
The 'easychanges' extension was added. It enhances Mercurial with a 'changes' command that allows to fire a graphical editor with the changes since the last revision (it is pretty much like the 'hg diff', expected that the whole file is displayed). EasyAPI was slightly enhanced. A bug was fixed in easycommmit when viewing added files.
[TAGS] Update, Feature, Fix